Remove PromoteTrigger using MQTT
