--- src/main/resources/hudson/plugins/accurev/AccurevPromoteTrigger/AccurevPromoteAction/index.jelly
@@ -1,22 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:l="/lib/layout"
->
-  <l:layout>
-    <st:include it="${it.owner}" page="sidepanel.jelly"/>
-    <l:main-panel>
-      <h1>${%Last Accurev Promote}</h1>
-      <j:set var="log" value="${it.log}"/>
-      <j:choose>
-        <j:when test="${empty(log)}">
-          ${%No promote has arrived yet.}
-        </j:when>
-        <j:otherwise>
-          <pre>
-            <st:getOutput var="output"/>
-            <j:whitespace>${it.writeLogTo(output)}</j:whitespace>
-          </pre>
-        </j:otherwise>
-      </j:choose>
-    </l:main-panel>
-  </l:layout>
-</j:jelly>
